requirements.txt
scikit-learn==1.3.2
matplotlib==3.7.4
apache-airflow-providers-mongo==3.5.0
apache-airflow-providers-oracle==3.9.1
clickhouse-connect==0.5.23
pandas==2.0.3
apache-airflow-providers-apache-spark